requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.13.0

# Data collection
requests>=2.31.0
gdelt>=0.1.10
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0
db-dtypes>=1.1.1  # Required for BigQuery pandas support

# Time series
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2

# Web scraping (for ground truth)
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Evaluation
scikit-metrics>=0.1.0
